+++ playground/src/main.ts
@@ -1,6 +1,17 @@
 import "./assets/main.css";
-
-import { createApp } from "vue";
+import { loadClientDirective} from "vue-onigiri/runtime/utils"
+import { createApp, Suspense, h } from "vue";
 import App from "./App.vue";
 
-createApp(App).mount("#app");
+createApp({
+    setup() {
+        return () =>  h(Suspense, null, {
+
+            default: () => h(App)
+        }
+        )
+    },
+    directives: {
+        'load-client': loadClientDirective
+    }
+}).mount("#app");