+++ playground/vite.config.ts
@@ -4,13 +4,13 @@ import { defineConfig } from "vite";
 import vueDevTools from "vite-plugin-vue-devtools";
 import { vueOnigiriPluginFactory } from "../src/vite/chunk";
 
-const { client, server } = vueOnigiriPluginFactory({
-  includeClientChunks: ["./src/components/HelloWorld.vue"],
+const { client } = vueOnigiriPluginFactory({
+  includeClientChunks: ["./src/components/Counter.vue", "./src/components/TsSlot.ts"],
 });
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [client(), vueDevTools(), server()],
+  plugins: [client(), vueDevTools()],
   resolve: {
     alias: {
       "@": fileURLToPath(new URL("src", import.meta.url)),