+++ src/vite/chunk.ts
@@ -1,20 +1,21 @@
 import type { Plugin } from "vite";
 import { hash } from "ohash";
 import MagicString from "magic-string";
-import type { ExportDefaultDeclaration } from "acorn";
-import { join, normalize, relative } from "node:path";
+ import { join, normalize, relative } from "node:path";
 import { readFileSync } from "node:fs";
 import vue from "@vitejs/plugin-vue";
 import { defu } from "defu";
 import type { Options } from "@vitejs/plugin-vue";
 import { glob } from "node:fs/promises";
+import type { ExportNamedDeclaration, ExportDefaultDeclaration, Declaration } from "estree";
+import type { AstNodeLocation, ProgramNode, RollupAstNode } from "rollup";
 
 function normalizePath(path: string): string {
   return normalize(path).replaceAll("\\", "/");
 }
 
 export type VSCOptions = {
-  includeClientChunks: string[];
+  includeClientChunks: (string | { path: string, export: string })[];
   /**
    * root directory from which to resolve the files.
    * @default {string} root of the project
@@ -25,10 +26,6 @@ export type VSCOptions = {
    */
   vueServerOptions?: Options;
   /**
-   * @default {string} build asset dir to store server chunks.
-   */
-  serverAssetsDir?: string;
-  /**
    * @default {string} build asset dir to store client chunks. Fallbacks to `build.assetsDir` if not provided.
    */
   clientAssetsDir?: string;
@@ -36,42 +33,24 @@ export type VSCOptions = {
 
 const VSC_PREFIX = "virtual:vsc:";
 const VSC_PREFIX_RE = /^(\/?@id\/)?virtual:vsc:/;
-const NOVSC_PREFIX_RE = /^(\/?@id\/)?(?!virtual:vsc:)/;
 
 export function vueOnigiriPluginFactory(options: Partial<VSCOptions> = {}): {
   client: (opts?: Options) => Plugin[];
   server: (opts?: Options) => Plugin[];
+  clientChunks: { originalPath: string; id: string; filename?: string; }[]
+  serverChunks: { originalPath: string; id: string; filename?: string; clientSideChunk?: string; serverChunkPath?: string }[]
 } {
-  const { serverAssetsDir = "", clientAssetsDir = "", rootDir = "" } = options;
-  const refs: { path: string; id: string }[] = [];
+  const { clientAssetsDir = "" } = options;
+  let { rootDir = "" } = options;
+  const clientChunks: { originalPath: string; id: string; filename?: string, exports: string[] }[] = [];
+  const serverChunks: { originalPath: string; id: string; filename?: string, exports: string[], clientSideChunk?: string, serverChunkPath?: string }[] = [];
   let assetDir: string = clientAssetsDir;
   let isProduction = false;
   return {
+    clientChunks: clientChunks,
+    serverChunks,
     client: (opts) => [
       vue(opts),
-
-      {
-        name: "vue-onigiri:renderSlotReplace",
-        transform: {
-          order: "post",
-          handler(code, id) {
-            if (VSC_PREFIX_RE.test(id)) {
-              const s = new MagicString(code);
-              s.prepend(
-                `import { renderSlot as cryoRenderSlot } from 'vue-onigiri/runtime/render-slot';\n`,
-              );
-
-              // replace renderSlot with vue-onigiri:renderSlot
-              s.replace(/_renderSlot\(/g, "cryoRenderSlot(_ctx,");
-
-              return {
-                code: s.toString(),
-                map: s.generateMap({ hires: true }).toString(),
-              };
-            }
-          },
-        },
-      },
       {
         name: "vite:vue-server-components-client",
         configResolved(config) {
@@ -80,50 +59,128 @@ export function vueOnigiriPluginFactory(options: Partial<VSCOptions> = {}): {
           }
           isProduction = config.isProduction;
           if (!rootDir) {
-            options.rootDir = config.root;
+            rootDir = config.root;
           }
         },
         async buildStart() {
-          const chunksToInclude = Array.isArray(options.includeClientChunks)
-            ? options.includeClientChunks
-            : [options.includeClientChunks || "**/*.vue"];
+          const chunksToInclude = Array.isArray(options.includeClientChunks) ? options.includeClientChunks : [options.includeClientChunks || "**/*.vue"];
 
-          const files = glob(chunksToInclude, {
-            cwd: rootDir,
-          });
-          for await (const file of files) {
-            const id = join(rootDir, file);
-            if (isProduction) {
-              const emitted = this.emitFile({
-                type: "chunk",
-                fileName: normalizePath(join(assetDir, hash(id) + ".mjs")),
-                id: id,
-                preserveSignature: "strict",
-              });
-              refs.push({
-                path: normalizePath(id),
-                id: normalizePath(this.getFileName(emitted)),
-              });
-            } else {
-              refs.push({
-                path: normalizePath(id),
-                id: normalizePath(join(clientAssetsDir, relative(rootDir, id))),
-              });
+          await Promise.all(chunksToInclude.map(async (file) => {
+            const path = typeof file === "string" ? file : file.path;
+            const exportName = typeof file === "string" ? 'default' : file.export;
+            const files = glob(path, {
+              cwd: rootDir,
+            });
+            for await (const file of files) {
+              const id = join(rootDir, file);
+
+              const info = clientChunks.find((chunk) => chunk.originalPath === normalizePath(id));
+              if (info) {
+                info.exports.push(exportName);
+              } else {
+                if (isProduction) {
+             
+                    clientChunks.push({
+                      originalPath: normalizePath(id),
+                      id: hash(id),
+                      exports: [exportName],
+                    });
+                } else {
+                  clientChunks.push({
+                    originalPath: normalizePath(id),
+                    id: normalizePath(join(clientAssetsDir, relative(rootDir, id))),
+                    exports: [exportName],
+                  });
+                }
+              }
+            }
+          }))
+        },
+        transform: {
+          order: 'post',
+          async handler(code, id) {
+            const shouldTransform = VSC_PREFIX_RE.test(id) || clientChunks.some((chunk) => chunk.originalPath === id);
+            if (!shouldTransform) {
+              return;
+            }
+
+            const ref = clientChunks.find(info => info.originalPath === normalizePath(id) || info.id === id);
+            if (!ref) {
+              return;
+            }
+
+            const s = new MagicString(code);
+            const ast = this.parse(code);
+
+           const exportNodes = ref.exports.map((exportName) => {
+              return [
+                exportName,
+                ast.body.find((node): node is ExportDefaultDeclaration | ExportNamedDeclaration => {
+                  if (exportName === "default") {
+                    return node.type === "ExportDefaultDeclaration";
+                  }
+                  return node.type === "ExportNamedDeclaration" && node.specifiers.some((specifier) => specifier.exported.type === "Identifier" && specifier.exported.name === exportName);
+                })
+              ] as const;
+            });
+            for (const [exportName, exportNode] of exportNodes) {
+              if (exportNode && exportNode.declaration) {
+                const { start, end } = exportNode.declaration as unknown as AstNodeLocation;
+                s.overwrite(
+                  start,
+                  end,
+                  `Object.assign(${code.slice(start, end)},
+                                    { __chunk: "${normalizePath(join("/", isProduction ? ref.id : relative(rootDir, normalize(ref.id))))}", __export: ${JSON.stringify(exportName)}  },
+                                )`
+                );
+              }
+            }
+            if (s.hasChanged()) {
+              return {
+                code: s.toString(),
+                map: s.generateMap({ hires: true }).toString(),
+              };
             }
           }
         },
+
 
         generateBundle(_, bundle) {
           for (const chunk of Object.values(bundle)) {
             if (chunk.type === "chunk") {
-              const list = refs.map((ref) => ref.id);
+              const list = clientChunks.values().map((ref) => ref.id).toArray();
               if (list.includes(chunk.fileName)) {
                 chunk.isEntry = false;
               }
             }
+          } 
+        }
+      },
+      {
+        name: 'load:vue-onigiri',
+        resolveId(id, importer, opts) {
+          if (VSC_PREFIX_RE.test(id) ) {
+            return this.resolve(id.replace(VSC_PREFIX, ""), importer?.replace(VSC_PREFIX_RE, ""), opts);
+          }
+          if(id === 'virtual:vue-onigiri') {
+            return id
           }
         },
-      },
+        load(id) {          
+            if(id === 'virtual:vue-onigiri') { 
+              return `
+              import { defineAsyncComponent } from "vue";
+              export default {
+                ${clientChunks.map(chunk =>
+                  chunk.exports.map(exportName =>
+                    `"${normalizePath(join("/", isProduction ? normalize(chunk.id) : relative(rootDir, normalize(chunk.id))))}#${exportName}": defineAsyncComponent(() => import("${chunk.originalPath}").then(m => m.${exportName}))`
+                  ).join(",\n")
+                ).join(",\n")}
+              }
+              `
+            }
+        },
+      }
     ],
 
     server: (opts) => [
@@ -133,30 +190,67 @@ export function vueOnigiriPluginFactory(options: Partial<VSCOptions> = {}): {
         enforce: "pre",
         name: "vite:vue-server-components-server",
         async buildStart() {
-          const chunksToInclude = Array.isArray(options.includeClientChunks)
-            ? options.includeClientChunks
-            : [options.includeClientChunks || "**/*.vue"];
-
-          const files = glob(chunksToInclude, {
-            cwd: rootDir,
-          });
-          for await (const file of files) {
-            const id = join(rootDir, file);
-            if (isProduction) {
-              this.emitFile({
-                type: "chunk",
-                fileName: normalizePath(
-                  join(serverAssetsDir, hash(id) + ".mjs"),
-                ),
-                id: id,
-                preserveSignature: "strict",
-              });
-            }
+          if (!isProduction) {
+            return
           }
+
+
+          if (options.includeClientChunks) {
+
+            await Promise.all((Array.isArray(options.includeClientChunks) ? options.includeClientChunks : [options.includeClientChunks]).map(async (file) => {
+              const path = typeof file === "string" ? file : file.path;
+              const exportName = typeof file === "string" ? 'default' : file.export;
+              const files = glob(path, {
+                cwd: rootDir,
+              });
+
+              for await (const file of files) {
+                const id = join(rootDir, file);
+                const info = serverChunks.find((chunk) => chunk.originalPath === normalizePath(id));
+                if (info) {
+                  info.exports.push(exportName);
+                } else {
+                  const clientSideChunk = clientChunks.find((chunk) => chunk.originalPath === normalizePath(id));
+
+                  if (isProduction) {
+                    const emitted = this.emitFile({
+                      type: "chunk",
+                      id: VSC_PREFIX + id,
+                      preserveSignature: "strict",
+                    });
+                    serverChunks.push({
+                      originalPath: normalizePath(id),
+                      id: emitted,
+                      exports: [exportName],
+                      clientSideChunk: clientSideChunk?.filename
+                    });
+                  } else {
+                    serverChunks.push({
+                      originalPath: normalizePath(id),
+                      id: normalizePath(join(clientAssetsDir, relative(rootDir, id))),
+                      exports: [exportName],
+                      clientSideChunk: clientSideChunk?.filename
+                    });
+                  }
+                }
+              }
+            }))
+          }
+
+          this.emitFile({
+            type: 'chunk',
+            fileName: 'vue-onigiri.mjs',
+            id: 'virtual:vue-onigiri',
+            preserveSignature: 'strict',
+          })
+
         },
         resolveId: {
           order: "pre",
-          async handler(id, importer) {
+          async handler(id, importer, opts) {
+              if (id === "virtual:vue-onigiri") {
+              return id;
+            }
             if (importer && VSC_PREFIX_RE.test(importer)) {
               if (VSC_PREFIX_RE.test(id)) {
                 return id;
@@ -165,21 +259,23 @@ export function vueOnigiriPluginFactory(options: Partial<VSCOptions> = {}): {
                 const resolved = await this.resolve(
                   id,
                   importer.replace(VSC_PREFIX_RE, ""),
+                  opts
                 );
                 if (resolved) {
                   return VSC_PREFIX + resolved.id;
                 }
               }
-              return this.resolve(id, importer.replace(VSC_PREFIX_RE, ""), {
-                skipSelf: true,
-              });
+              return this.resolve(id, importer.replace(VSC_PREFIX_RE, ""), opts);
             }
-
+            if (id.endsWith(".ts") || id.endsWith(".tsx") || id.endsWith(".js") || id.endsWith(".jsx")) {
+              return this.resolve(id.replace(VSC_PREFIX_RE, ""), importer?.replace(VSC_PREFIX_RE, ""), opts);
+            }
             if (VSC_PREFIX_RE.test(id)) {
               if (id.replace(VSC_PREFIX_RE, "").startsWith("./")) {
                 const resolved = await this.resolve(
                   id.replace(VSC_PREFIX_RE, ""),
                   importer?.replace(VSC_PREFIX_RE, ""),
+                  opts
                 );
                 if (resolved) {
                   return VSC_PREFIX + resolved?.id;
@@ -187,100 +283,91 @@ export function vueOnigiriPluginFactory(options: Partial<VSCOptions> = {}): {
               }
               return id;
             }
-          },
+          }
         },
         load: {
           order: "pre",
           async handler(id) {
             const [filename, rawQuery] = id.split(`?`, 2);
+            if (!rawQuery && VSC_PREFIX_RE.test(id) && id.endsWith('.vue')) {
+              const file = id.replace(VSC_PREFIX_RE, "");
 
-            if (!rawQuery) {
-              if (VSC_PREFIX_RE.test(id)) {
-                const file = id.replace(VSC_PREFIX_RE, "");
-
-                return {
-                  code: readFileSync(normalizePath(normalize(file)), "utf8"),
-                };
-              }
-              if (filename?.endsWith(".vue")) {
-                const fileName = join(serverAssetsDir, hash(id) + ".mjs");
-                this.emitFile({
-                  type: "chunk",
-                  fileName,
-                  id: VSC_PREFIX + id,
-                  preserveSignature: "strict",
-                });
-              }
-            }
-          },
-        },
-
-        generateBundle(_, bundle) {
-          for (const chunk of Object.values(bundle)) {
-            if (chunk.type === "chunk") {
-              const list = refs.map((ref) => ref.id);
-              if (list.includes(chunk.fileName)) {
-                chunk.isEntry = false;
-              }
-            }
-          }
-        },
-
-        transform: {
-          order: "post",
-          handler(code, id) {
-            const ref = refs.find(
-              (ref) => ref.path === id.replace(VSC_PREFIX_RE, ""),
-            );
-
-            if (ref) {
-              const s = new MagicString(code);
-              const ast = this.parse(code);
-              const exportDefault = ast.body.find((node) => {
-                return node.type === "ExportDefaultDeclaration";
-              }) as
-                | (ExportDefaultDeclaration & { start: number; end: number })
-                | undefined;
-              const ExportDefaultDeclaration = exportDefault?.declaration;
-              if (ExportDefaultDeclaration) {
-                const { start, end } = ExportDefaultDeclaration;
-                s.overwrite(
-                  start,
-                  end,
-                  `Object.assign(
-                                    { __chunk: "${normalizePath(join("/", isProduction ? normalize(ref.id) : relative(rootDir, normalize(ref.id))))}" },
-                                     ${code.slice(start, end)},
-                                )`,
-                );
-                return {
-                  code: s.toString(),
-                  map: s.generateMap({ hires: true }).toString(),
-                };
-              }
-            }
-          },
-        },
-      },
-      {
-        name: "vue-onigiri:renderSSRSlotReplace",
-        transform: {
-          order: "post",
-          handler(code, id) {
-            if (VSC_PREFIX_RE.test(id)) {
-              const s = new MagicString(code);
-              s.prepend(
-                `import { renderSlot as cryoRenderSlot } from 'vue-onigiri/runtime/render-slot';\n`,
-              );
-              // replace renderSlot with vue-onigiri:renderSlot
-              s.replace(/_renderSlot\(/g, "cryoRenderSlot(_ctx,");
               return {
-                code: s.toString(),
-                map: s.generateMap({ hires: true }).toString(),
+                code: readFileSync(normalizePath(normalize(file)), "utf8"),
               };
             }
           },
         },
+
+        
+        generateBundle(output, bundle) {
+          for (const chunk of Object.values(bundle)) {
+            if (chunk.type === "chunk") {
+              const list = serverChunks.values().map((ref) => ref.id).toArray();
+              if (list.includes(chunk.fileName)) {
+                chunk.isEntry = false;
+                chunk.isImplicitEntry = false;
+                chunk.isDynamicEntry = false;
+              }
+            }
+          } 
+        },
+
+        transform: {
+          order: "post",
+          handler(code, id) {
+               const ref = clientChunks.find((chunk) => chunk.originalPath === id.replace(VSC_PREFIX_RE, ""));
+
+            if (id && ref) {
+              const s = new MagicString(code);
+              const ast = this.parse(code) as RollupAstNode<ProgramNode>;
+              for (const exportName of ref.exports) {
+                const exportNode = ast.body.find((node): node is (ExportNamedDeclaration | ExportDefaultDeclaration) => {
+                  if (exportName === "default") {
+                    return node.type === "ExportDefaultDeclaration";
+                  }
+                  return node.type === "ExportNamedDeclaration" && node.specifiers.some((specifier) => specifier.exported.type === "Identifier" && specifier.exported.name === exportName);
+                });
+                if (exportNode) {
+                                  if (exportNode.declaration) {
+                
+                                  const { start, end } = exportNode.declaration as unknown as RollupAstNode<Declaration>; 
+                                    s.overwrite(
+                                      start,
+                                      end,
+                                      `Object.assign( ${code.slice(start, end)},
+                                                      { __chunk: "${normalizePath(join("/", isProduction ? ref.id : relative(rootDir, normalize(id))))}", __export: ${JSON.stringify(exportName)} },
+                                                      
+                                                  )`,
+                                    );
+                                  } else { 
+                                    // todo
+                                  }
+                                }
+              }
+              if (s.hasChanged()) {
+                return {
+                  code: s.toString(),
+                  map: s.generateMap({ hires: true }).toString()
+                };
+              }
+            }
+          },
+        },
       },
+       
+      {
+        name: 'vue:onigiri:loadvirtual',
+        load(id) {
+          if (id === "virtual:vue-onigiri") {
+            return `
+              ${serverChunks.map((chunk, index) => `import * as i${index} from '${VSC_PREFIX + chunk.originalPath}'`).join("\n")}
+              export default new Map( [
+            ${serverChunks.map((chunk, index) => `[${JSON.stringify("/" + chunk.clientSideChunk)}, i${index}]`).join(",\n")}
+          ] );`;
+          }
+        }
+      }
     ],
   };
 }
@@ -296,36 +383,31 @@ function getVuePlugin(options?: Options) {
   );
   // need to force non-ssr transform to always render vnode
   const oldTransform = plugin.transform;
-  plugin.transform = async function (code, id, _options) {
+  plugin.transform = async function (code, id, options) {
     if (VSC_PREFIX_RE.test(id)) {
       return;
     }
     // @ts-expect-error blabla
-    return await Reflect.apply(oldTransform, this, [code, id, { ssr: false }]);
+    return await Reflect.apply(oldTransform, this, [code, id, options]);
   };
   const oldLoad = plugin.load;
-  plugin.load = async function (id, _options) {
+  plugin.load = async function (id, options) {
     if (VSC_PREFIX_RE.test(id)) {
       return;
     }
     // @ts-expect-error blabla
-    return await Reflect.apply(oldLoad, this, [id, { ssr: false }]);
+    return await Reflect.apply(oldLoad, this, [id, options]);
   };
 
   return plugin;
 }
 
 function getPatchedServerVue(options?: Options): Plugin {
-  const plugin = vue(
-    defu(options, {
-      include: [VSC_PREFIX_RE],
-      exclude: [NOVSC_PREFIX_RE],
-    }),
-  );
+  const plugin = vue(defu(options));
   // need to force non-ssr transform to always render vnode
   const oldTransform = plugin.transform;
   plugin.transform = async function (code, id, _options) {
-    if (!id.includes(".vue")) {
+    if (!VSC_PREFIX_RE.test(id)) {
       return;
     }
     // @ts-expect-error blabla
@@ -333,7 +415,7 @@ function getPatchedServerVue(options?: Options): Plugin {
   };
   const oldLoad = plugin.load;
   plugin.load = async function (id, _options) {
-    if (!id.includes(".vue")) {
+    if (!VSC_PREFIX_RE.test(id)) {
       return;
     }
     // @ts-expect-error blabla