+++ test/serialization.test.ts
@@ -124,6 +124,7 @@ describe("serialize/deserialize", () => {
               1,
               undefined,
               "/test/fixtures/components/Counter.vue",
+              "default",
               {
                 "default": [
                   [
@@ -357,7 +358,48 @@ describe("slots", () => {
     `);
     const astHtml = await renderToString(renderOnigiri(ast)!);
     expect(removeCommentsFromHtml(astHtml)).toMatchInlineSnapshot(
-      `"<div><div> counter : 0 <button>Increment</button><div><p>Slot to Counter: 0</p></div></div></div>"`,
+      `
+      [
+        0,
+        "div",
+        undefined,
+        [
+          [
+            1,
+            undefined,
+            "/test/fixtures/components/Counter.vue",
+            "default",
+            {
+              "default": [
+                [
+                  3,
+                  [
+                    [
+                      0,
+                      "div",
+                      undefined,
+                      [
+                        [
+                          0,
+                          "p",
+                          undefined,
+                          [
+                            [
+                              2,
+                              "Slot to Counter: 0",
+                            ],
+                          ],
+                        ],
+                      ],
+                    ],
+                  ],
+                ],
+              ],
+            },
+          ],
+        ],
+      ]
+    `,
     );
     const html = await renderToString(h(SlotToCounter));
     expect(removeCommentsFromHtml(html)).toMatchInlineSnapshot(